timescaledb/requirements.txt
psycopg2==2.9.6
python-dateutil==2.8.2
pandas==2.1.4
openpyxl==3.1.2
